Add abstract class for the problem, which contains common methods like the volume correction and computation of the weighted volume
